feat: simplify deployNitroMocks API and add hardhat deployment task
  - Add hardhat task for deploying precompiles to standalone nodes
  - Remove ethers/network/signer parameters from deployNitroMocks()
  - Support incremental deployment of specific precompiles
  - Auto-detect hardhat environment instead of requiring parameters
  - Consolidate documentation into main README usage section

+++ test/arbowner/set-brotli-compression-level.test.ts
@@ -22,7 +22,7 @@ describe("ArbOwner.setBrotliCompressionLevel", function () {
       {
         storageAccess: storageAccessComparerExcludingVersion,
         result: (mock, _underlying) => {
-          originalLevel = mock;
+          originalLevel = mock as bigint;
         }
       }
     );
